Attach names and annotations to method parameters
Instead of having just list of types, JavaMethod now has a list of
JavaParameters that contains names (if available in class file) and
annotations for the parameter.

Due to the structure of the class file loading, parameter list is first
created based on available types and later on the JavaParameter objects
are augmented with additional details.