Separate type schema from variables
Instead of making type-variables know their bounds, separate the schema
and the variables. The variables themselves are just keys to the schema,
which knows the bounds.

The change is necessary to parse anonymous inner classes occurring inside
generic methods, because variables of the method are not available when
parsing the schema.