Generate code to in-memory buffer instead of streaming
The previous TypeScript generation code is unnecessary complex because
outputs to arbitrary Appendable, requiring IOExceptions to be handled
all over the place and producing other constraints.

Since the amount of memory required to store the whole generated code
is really small anyway, this commit simplifies the code by just storing
everything into a StringBuilder. This makes both the writer and the
using code simpler.

This change also has the added advantage that the code generator is not
blocking on I/O so it could possibly improve the total performance.