Resolve all JavaTypes against environment
Previous code just looked if JavaType is a variable and resolved that
against environment if it was, otherwise returning the type as it is.
This fails with more complex types, e.g. if the type is List<T> and
we know that T=String. New code resolves the whole type hierarchy
recursively.